requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
torch==2.3.1
torchvision==0.18.1
Pillow==10.4.0
pydantic==2.9.2
